requirements.txt
colorful==0.5.4
Faker==13.3.4
  python-dateutil==2.8.2
    six==1.16.0
inquirerpy==0.3.3
  pfzy==0.3.4
  prompt-toolkit==3.0.29
    wcwidth==0.2.5
peewee==3.14.10
phonenumberslite==8.12.46
setuptools==44.0.0
smart-getenv==1.1.0
wheel==0.37.1
